+++ scripts/merchants.js
@@ -11,9 +11,16 @@ let merchants = {
       {item: copy(weapons.battle_axe), price: 950, type: "weapon"},
       {item: copy(weapons.long_sword), price: 365, type: "weapon"},
       {item: copy(weapons.dual_daggers), price: 720, type: "weapon"},
+      {item: copy(weapons.poleaxe), price: 1895, type: "weapon"},
+      {item: copy(weapons.enchanted_blade), price: 3000, type: "weapon"},
       {item: copy(armors.rags), price: 50, type: "armor"},
       {item: copy(armors.leather_armor), price: 280, type: "armor"},
       {item: copy(armors.chainmail), price: 310, type: "armor"},
+      {item: copy(armors.enchanted_robes), price: 855, type: "armor"}
     ]
+  },
+  floor2_merchant: {
+    name: "Maruda Baneforse",
+    stock: [],
   }
 }